Match unsigned collision triangle traversal
Use unsigned triangle counts and indices so both collision-list loops emit the retail cmplw guards.

Constraint: Preserve the recovered duplicate-height control flow.
Rejected: Signed loop counters | diverge from target and mis-handle high-bit counts.
Confidence: high
Scope-risk: narrow
Tested: Forced gs_range_8010CBD0 build; fn_8010E138 51.85% to 52.21%; git diff --check.

+++ src/game/gs_range_8010CBD0.c
@@ -812,8 +812,8 @@ s32 fn_8010E138(void* origin, void* direction) {
     u8 field31;
     s32 cellX;
     s32 cellZ;
-    s32 triCount;
-    s32 triIdx;
+    u32 triCount;
+    u32 triIdx;
 
     point = origin;
     outHit = direction;
@@ -843,7 +843,7 @@ s32 fn_8010E138(void* origin, void* direction) {
             list = (GScolsys2TriangleList*)region->triList;
             triangle = list->triangles;
             for (triIdx = 0;
-                 triIdx < (s32)list->count && temporaryCount < 8;
+                 triIdx < list->count && temporaryCount < 8;
                  triIdx++, triangle++)
             {
                 for (i = 0; i < 3; i++) {
